RF | Move tests from the FooterTest2 class to the FooterTest class
