+++ src/test/java/school/redrover/FooterTest.java
@@ -1,16 +1,20 @@
 package school.redrover;
 
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.testng.Assert;
-import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.runner.BaseTest;
 
 public class FooterTest extends BaseTest {
+
+    private static final By JENKINS_VERSION_BTN = By.xpath("//a[@rel='noopener noreferrer']");
+    private static final By MANAGE_JENKINS_BTN = By.xpath("//span[contains(text(),'Manage Jenkins')]/..");
+    private static final String JENKINS_VERSION = "Jenkins 2.387.2";
+    private static final By REST_API_BTN = By.xpath("//a[contains(text(),'REST API')]");
+    private static final String TITLE = "Remote API [Jenkins]";
 
     @Test
     public void testRestAPILink() {
@@ -45,4 +49,24 @@ public class FooterTest extends BaseTest {
 
         Assert.assertEquals(getDriver().findElement(By.xpath("//a[@rel='noopener noreferrer']")).getText(),"Jenkins 2.387.2");
     }
+
+    @Test
+    public void testVerifyJenkinsVersionOnManageJenkinsPage() {
+        getDriver().findElement(MANAGE_JENKINS_BTN).click();
+
+        Actions actions = new Actions(getDriver());
+        actions.scrollToElement(getDriver().findElement(JENKINS_VERSION_BTN));
+
+        Assert.assertEquals(getDriver().findElement(JENKINS_VERSION_BTN).getText(), JENKINS_VERSION, "Wrong version Jenkins");
+    }
+
+    @Test
+    public void testVerifyRESTAPILink(){
+        Actions actions = new Actions(getDriver());
+        actions.scrollToElement(getDriver().findElement(REST_API_BTN));
+
+        getDriver().findElement(REST_API_BTN).click();
+        Assert.assertEquals(getDriver().getTitle(),TITLE, "Wrong page or page title");
+    }
+
 }
